# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Yousra Stock API",
    "version": "1.0.0",
    "description": "Internal ERP/stock management API for Yousra Company.\n\n**Response envelope**\n- Single resource: `{ \"data\": { ... } }`\n- List: `{ \"data\": [ ... ], \"meta\": { \"page\", \"pageSize\", \"total\" } }`\n- Error: `{ \"error\": { \"message\": string, \"code\": string } }`"
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiTokenAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Token",
        "description": "Opaque token issued when an external system is registered. Service-to-service only."
      }
    },
...
    "schemas": {
...
      "ExternalConsumeInput": {
        "type": "object",
        "properties": {
          "system": {
            "type": "string",
            "minLength": 1
          },
          "orderRef": {
            "type": "string",
            "minLength": 1
          },
          "locationCode": {
            "type": "string",
            "minLength": 1
          },
          "lines": {
            "minItems": 1,
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "itemId": {
                  "type": "string",
                  "format": "uuid"
                },
                "quantity": {
                  "type": "number",
                  "minimum": 0,
                  "exclusiveMinimum": true
                }
              },
              "required": [
                "itemId",
                "quantity"
              ]
            }
          }
        },
        "required": [
          "system",
          "orderRef",
          "locationCode",
          "lines"
        ]
      },
      "ExternalReleaseInput": {
        "type": "object",
        "properties": {
          "system": {
            "type": "string",
            "minLength": 1
          },
          "orderRef": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "system",
          "orderRef"
        ]
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "message",
              "code"
            ],
            "properties": {
              "message": {
                "type": "string",
                "example": "Resource not found"
              },
              "code": {
                "type": "string",
                "example": "NOT_FOUND",
                "description": "Machine-readable code: NOT_FOUND, VALIDATION_ERROR, UNAUTHORIZED, FORBIDDEN, CONFLICT, INTERNAL_ERROR."
              },
              "details": {
                "type": "object",
                "nullable": true,
                "description": "Only on VALIDATION_ERROR — Zod `flatten()` output."
              }
            }
          }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Validation failed",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": {
                "message": "Validation failed",
                "code": "VALIDATION_ERROR"
              }
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing or invalid bearer token",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": {
                "message": "Authentication required",
                "code": "UNAUTHORIZED"
              }
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": {
                "message": "Resource not found",
                "code": "NOT_FOUND"
              }
            }
          }
        }
      },
      "Conflict": {
        "description": "Conflicts with current state",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": {
                "message": "Insufficient stock for this movement",
                "code": "CONFLICT"
              }
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/external/stock": {
      "get": {
        "tags": [
          "External"
        ],
        "summary": "Current stock for a set of items at one location (batch)",
        "description": "`companyId` is derived from the API token. `itemIds` are Yousra Item UUIDs; an item with no movements yet reports quantity `\"0\"`.\n",
        "security": [
          {
            "apiTokenAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "system",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Caller's registered name; must match the token."
          },
          {
            "in": "query",
            "name": "locationCode",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Warehouse code (e.g. akwa-branch)."
          },
          {
            "in": "query",
            "name": "itemIds",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Comma-separated Yousra Item UUIDs (max 100)."
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "locationCode": {
                          "type": "string"
                        },
                        "warehouseId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "items": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "itemId": {
                                "type": "string",
                                "format": "uuid"
                              },
                              "quantity": {
                                "type": "string",
                                "example": "42"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/external/stock/consume": {
      "post": {
        "tags": [
          "External"
        ],
        "summary": "An external order was confirmed — decrement stock (one SALE movement per line)",
        "description": "Idempotent on `orderRef`: a repeat returns the original movements with status 200 and moves no stock. Insufficient stock for any line rolls the whole call back with a 409. Movements are attributed to the calling external system, not a user.\n",
        "security": [
          {
            "apiTokenAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExternalConsumeInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Replay — original movements returned",
            "no change": null
          },
          "201": {
            "description": "Stock decremented"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      }
    },
    "/external/stock/release": {
      "post": {
        "tags": [
          "External"
        ],
        "summary": "A previously-consumed external order was cancelled — restore its stock",
        "description": "Reverses every SALE movement recorded for `orderRef` with a matching RETURN movement. Idempotent: a repeat returns the original RETURNs with status 200. 404 if no consumption was ever recorded for the order.\n",
        "security": [
          {
            "apiTokenAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExternalReleaseInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Replay — original RETURN movements returned",
            "no change": null
          },
          "201": {
            "description": "Stock restored"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
...
}